fix: SM-177 useAxios Hooks 위반 문제 해결
