+++ src/utils/apis/cocktail.ts
@@ -1,5 +1,5 @@
 import useAxios from '@hooks/useAxios';
-import type { ICocktail } from '@models/types';
+import type { ICocktail, ITHEME } from '@models/types';
 import { AXIOS_REQUEST_TYPE } from '@constants/axios';
 
 const searchCocktailByName = (keyword: string): Promise<ICocktail[]> => {
@@ -7,4 +7,14 @@ const searchCocktailByName = (keyword: string): Promise<ICocktail[]> => {
   return request.get(`/cocktail/${keyword}`);
 };
 
-export { searchCocktailByName };
+const getCocktailListByTheme = (
+  mainCategory: ITHEME,
+  subCategory: string
+): Promise<ICocktail[]> => {
+  const request = useAxios(AXIOS_REQUEST_TYPE.DEFAULT);
+  return request.get(
+    `/cocktail/theme?mainCategory=${mainCategory}&subCategory=${subCategory}`
+  );
+};
+
+export { searchCocktailByName, getCocktailListByTheme };
