+++ src/utils/apis/user.ts
@@ -1,7 +1,6 @@
 import useAxios from '@hooks/useAxios';
-import type { IIngredient, IUser } from '@models/types';
+import type { IIngredient } from '@models/types';
 import { AXIOS_REQUEST_TYPE } from '@constants/axios';
-import type { IRegisterRequestBody } from '@pages/RegisterPage/types';
 
 const updateUserIngredients = (
   userId: string,
@@ -11,18 +10,4 @@ const updateUserIngredients = (
   return request.post(`/user/ingredient/${userId}`, selectedIngredients);
 };
 
-const checkUserNickname = (
-  nickname: string
-): Promise<{ data: { can: boolean } }> => {
-  const request = useAxios(AXIOS_REQUEST_TYPE.DEFAULT);
-  return request.get(`/user/nickname/${nickname}`);
-};
-
-const registerUser = (
-  data: IRegisterRequestBody
-): Promise<IUser | undefined> => {
-  const request = useAxios(AXIOS_REQUEST_TYPE.DEFAULT);
-  return request.post('/user/join', data);
-};
-
-export { updateUserIngredients, checkUserNickname, registerUser };
+export { updateUserIngredients };
